Julia: split symbolic-node.jl into several snippets
- `symbolic-node/type.jl`
- `symbolic-node/show.jl`
- `symbolic-node/arithmetic.jl`
- `symbolic-node/io.jl`
- `symbolic-node/array.jl`
- `symbolic-node/op.jl`
- `symbolic-node/autodiff.jl`

See also: #14001
